gh-108303: Move tokenize-related data to Lib/test/tokenizedata (GH-109265)
(cherry picked from commit 1110c5bc828218086f6397ec05a9312fb73ea30a)
